Add files via upload
This is the current home page of the Optima website. The file named CSS/style.css is the style sheet for particle js and the background settings of logos, and social media icons and the footer. The js file named as js/index.js is the js file for particle js. Access these files directly from your page for styling footer and home page.